- floatified playerisdead

+++ source/games/duke/src/funct.h
@@ -115,7 +115,7 @@ int wakeup(DDukeActor* sn, int pn);
 int timedexit(int snum);
 void dokneeattack(int snum, const std::initializer_list<int>& respawnlist);
 int endoflevel(int snum);
-void playerisdead(int snum, int psectlotag, int fz, int cz);
+void playerisdead(int snum, int psectlotag, double fz, double cz);
 void footprints(int snum);
 int makepainsounds(int snum, int type);
 void playerCrouch(int snum);
